Add User Registration & Login Routes
Add a hashed password & auth token field to the Users table. The
password is hashed using bcrypt, and the token is a UUID generated with
a cryptographically secure algorithm.

Replace the generic CRUD routes for users with a registration & login
route. These routes return all the users data except for their hashed
password.

Add an authorization combinator allowing clients to pass a Token using
the `Auth-Token` HTTP Header. This combinator is not yet used and no
Routes have been modified to handle the Token.