Require Authentication for Create Routes
Add the GuardCRUD typeclass to verify that a User has the proper
permissions for a Resource's Create route. Eventually this will be
reused for the List, Read, Update, & Delete routes as well.

Authentication is required for all Create routes by default. Only
Administrators are allowed to create Routines & only Routine Authors are
allowed to create Sections & Section Exercises for a Routine.