Handle Authentication for List Routes
Add the `guardList` function to the `GuardCRUD` typeclass to generate a
list of Filters to apply the `selectList` operation. The default
implementation applies no filtering, returning all instances of the
Resource.

The Routine implementation allows anonymous users to see all public
routines, authorized users to see their private routines as well, and
administrators to see all routines. The Section & SectionExercise
implementations defer to the Routine implementation.